fix:  rename back file for diffs
